backend/requirements.txt
asgiref==3.6.0
Django==4.2
djangorestframework==3.14.0
Pillow==9.5.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3